acc dist visualizations added
the idea is that we would try to avoid classifying data that is
ambiguous somehow, by using a cutoff value to be calculated and defined.

what we saw from the graphs is that some classifiers fail do distinguish
the data at all, i. e., do not converge into something that classifies
them, so we need to make sure that at least at some point the data is
classified, or it will not work.